# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BRAMON API",
        "contact": {
            "name": "Support Team",
            "email": "support@bramonmeteor.org"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/admin/captures": {
            "get": {
                "tags": [
                    "Administrators"
                ],
                "summary": "List all captures",
                "operationId": "/v1/admin/captures",
                "responses": {
                    "200": {
                        "description": "List all captures",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "post": {
                "tags": [
                    "Administrators"
                ],
                "summary": "Upload files and create a capture register",
                "operationId": "/v1/admin/captures",
                "parameters": [
                    {
                        "name": "files[]",
                        "in": "path",
                        "description": "The capture files",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": "string"
                        }
                    },
                    {
                        "name": "station_id",
                        "in": "path",
                        "description": "The station identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The user identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the capture data.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When capture not exists or not public."
                    },
                    "422": {
                        "description": "Error: Bad Param. When file not found or can not be deleted."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Administrators"
                ],
                "summary": "Exclude captures files",
                "operationId": "/v1/admin/captures",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "The capture filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Returns a empty body when success.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When capture not exists or not public."
                    },
                    "422": {
                        "description": "Error: Bad Param. When file not found or can not be deleted."
                    }
                }
            }
        },
        "/v1/admin/captures/{capture}": {
            "get": {
                "tags": [
                    "Administrators"
                ],
                "summary": "View a capture",
                "operationId": "/v1/admin/captures/000-000-0000",
                "parameters": [
                    {
                        "name": "capture",
                        "in": "path",
                        "description": "The capture identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the capture details.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When capture not exists or not public."
                    }
                }
            }
        },
...
    }
}